extract_images_from_pdf/requirements.txt
pdf2image>=1.17.0
pillow>=10.0.0
